+++ src/sistemahotel/model/Hospede.java
@@ -8,10 +8,7 @@ package sistemahotel.model;
  *
  * @author klayt
  */
-public class Hospede extends Funcionario {
-    private int id;
-    private String nome;
-    private String cpf;
+public class Hospede extends Pessoa {
 
     public Hospede() {}
 
@@ -20,37 +17,7 @@ public class Hospede extends Funcionario {
         this.nome = nome;
         this.cpf = cpf;
     }
-
-    @Override
-    public int getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Override
-    public String getNome() {
-        return nome;
-    }
-
-    @Override
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    @Override
-    public String getCpf() {
-        return cpf;
-    }
-
-    @Override
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-    
+   
     @Override
     public String toString() {
         // O JComboBox vai mostrar o nome e o CPF do hóspede
